LayoutTests/platform/glib/http/tests/contentextensions/block-private-click-measurement-expected.txt

CONSOLE MESSAGE: Blocked by content extension
CONSOLE MESSAGE: Cannot load image https://127.0.0.1:8443/privateClickMeasurement/resources/redirectToConversion.py?conversionData=12 due to access control checks.
Tests that private click measurement redirects to well-known location don't trigger a conversion if they are blocked by content blockers.



--------
Frame: '<!--frame1-->'
--------
Attribution not received - timed out.

Unattributed Private Click Measurements:
WebCore::PrivateClickMeasurement 1
Source site: 127.0.0.1
Attribute on site: localhost
Source ID: 3
No attribution trigger data.
Application bundle identifier: com.apple.mobilesafari
